clang format and tidy

+++ speed-dreams/source-2.2.3/src/drivers/assistedhuman/Driver.cpp
@@ -43,9 +43,9 @@ void Driver::InitTrack(tTrack* p_track, void* p_carHandle, void** p_carParmHandl
     SMediator::GetInstance()->RaceStart(p_track, p_carHandle, p_carParmHandle, p_situation, m_recorder);
     auto end = std::chrono::system_clock::now();
 
-    long milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count() / 5;
+    long milliseconds = static_cast<long>(std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count() / 5);
 
-    if(milliseconds > 1000)
+    if (milliseconds > 1000)
     {
         GfLogWarning("SLOW BLACKBOX: Blackbox took %ld milliseconds (on average) to respond over %i tests, this is relatively slow!\n", milliseconds, BLACK_BOX_TESTS);
         return;
